divide elements.py in multiple modules
