+++ weconnect/elements/control_operation.py
@@ -1,0 +1,15 @@
+from enum import Enum
+
+
+class ControlInputEnum(Enum):
+    @classmethod
+    def allowedValues(cls):
+        return []
+
+
+class ControlOperation(ControlInputEnum):
+    START = 'start'
+    STOP = 'stop'
+    NONE = 'none'
+    SETTINGS = 'settings'
+    UNKNOWN = 'unknown'
